refactor: Rename main functions for clarity
Rename the main functions in benchmark and regression test files to
provide better context.

Update the package declaration to 'benchmarks' and adjust the script
for running benchmarks and regression tests to point to the new main
file locations.

- Change package name from 'main' to 'benchmarks'
- Rename 'main' functions to 'RunBenchmarkSuite' and 'RunRegressionTestSuite'
- Update run script to reflect new main file paths
